+++ src/User/Concerns/LegacyConstants.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CraftCms\Cms\User\Concerns;
+
+if (! class_exists(LegacyConstants::class)) {
+    /**
+     * @internal
+     */
+    #[\Deprecated(message: 'This is a hook for the yii2-adapter to define its constants.', since: '6.0.0')]
+    trait LegacyConstants {}
+}
